Add crash reporter
The crash reporter is a reporter that is suited for test suites
that can crash (segmentation fault and alike) and when multiple
bandit invokations are run in parallel.

When a test item is started, the crash reporter prints the
whole context (separated by " | ") and the item on one line.

The test summary only contains failures and is prefixed by
'#' so that you can grep for it. For example, using the default
failure formatter,

  run_tests --reporter=crash >log 2>&1 ||
     grep -B 1 -A 4 "^#" log ||
     tail -n 10 log

gives a nice output of the failure summary (if there are failures)
or outputs the last 10 lines of the output (if it crashes)
or is just silent otherwise.

Resolves #83 where such a reporter has been requested.